+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Now create README.md in meta directory to explain how to use it.
 - [#2](https://github.com/creekorful/srcode/issues/2) Implement srcode ls.
+- [#3](https://github.com/creekorful/srcode/issues/3) Implement srcode bulk-git.
 
 ## Changed